requirements.txt
attrs==25.4.0
cattrs==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
flatbuffers==25.9.23
future==1.0.0
geocoder==1.38.1
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
gTTS==2.5.4
h11==0.16.0
httplib2==0.31.2
idna==3.11
jh2==5.0.10
niquests==3.17.0
numpy==2.4.2
oauthlib==3.3.1
openmeteo_requests==1.7.5
openmeteo_sdk==1.25.0
pandas==3.0.1
platformdirs==4.9.2
playsound==1.2.2
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
qh3==1.5.6
ratelim==0.1.6
requests==2.32.5
requests-cache==1.3.0
requests-oauthlib==2.0.0
retry-requests==2.0.0
rsa==4.9.1
six==1.17.0
SpeechRecognition==3.14.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.16.900
volume-control==0.9.2
wassima==2.0.5
